fix(flamegraph): Paint artifacts on dirty region repaint
This was due to an incorrect assumption that the clip bounds of a nested
graphical context created like this

g2.create(rectOnSurface.x, rectOnSurface.y, rectOnSurface.width, rectOnSurface.height)

were always the same. It isn't in particular on repaint of a dirty
region.

Fixing this was as simple as passing the correct shape. In the paint
method. More over this led to simplifying some code in that area by
avoiding the mess with the clip.

Fixes #15