fix(flamegraph): Scrollpane now correctly behaves in SWT
Previously when the component was added to an SWT embedded
composite, the scrollbar weren't shown and the scroll pane
didn't scroll, a user had to resize the SWT component so that
JScrollPane behaves correctly.

It appears the main cause is related to the use of
`invalidate()` rather than `revalidate()` on the canvas.
This is now fixed.

Fixes #79