fix(flamegraph): Maintain the zoom factor when container shrinks
Custom layout code to handle container shrinking.
The default view port layout asks the preferred size
of the view.

But that cannot work since the canvas won;t update
its width, it receives its size from the layout container.
However, the default algorithm only updates the size
after it has received the preferred size, or if the
viewport got bigger.

This code makes the necessary query to the canvas to
asks if it needs a new size given the viewport width change,
in order to keep the same zoom factor.

The view location is also updated.

Fixes #46