fix: Rare case, in Flamegraph mode when resize could enter a loop
Computes the dimension as if a vertical scrollbar was needed.

Otherwise, the layout can enter a loop:
Because the view port width is called once with full width,
which computes a canvas with a taller dimension than viewport.
This triggers the horizontal scrollbar to be added, which
triggers another layout.
In this layout, the view port width is shorter by the scrollbar width,
which makes the canvas fitting in the view port,
which then triggers annoter layout that removes the vertical scrollbar,
and then starts again.

Note the scrollbar visibility is updated at the end of this control block