fix: correct frame alignment in Flamegraph, in mode switch & expand/shrink
Removed the Horizontal ScrollBar (HSB) from the ScrollPane's column
header, because it shifts the `y` view position by the HSB's height.
This requires propagating this shift in various places.
Also the ColumnHeader visibility needs to be tied to the HSB, otherwise
it remains always visible.

For this reason a simpler approach is taken by always showing the HSB
in Flamegraph mode.