fix: Downgrade javier semver plugin to 0.8.0 due to kotlin metadata incompatibility
This avoids warnings in the build that were introduced
as a workaround (using kotlin 2.2.0) with gradle semver 0.9.0

For reference this was failing with the following message:

Module was compiled with an incompatible version of Kotlin. The binary version of its metadata is 2.2.0, expected version is 2.0.0.

Note that this raise the warning :
WARNING: Unsupported Kotlin plugin version.
The `embedded-kotlin` and `kotlin-dsl` plugins rely on features of Kotlin `2.0.21` that might work differently than in the requested version `2.2.0`.